third_party/blink/web_tests/external/wpt/svg/animations/animateMotion-ellipse-expected.txt

This is a testharness.js-based test.
[FAIL] Test for checking position of the svg element when animateMotion's mpath is an ellipse
  assert_approx_equals: expected -200 +/- 1 but got 0
Harness: the test ran to completion.
